+++ .theia/settings.json
@@ -1,3 +1,3 @@
 {
-    "editor.autoSave": "on"
+    "editor.autoSave": "off"
 }
