+++ public/index.html
@@ -39,6 +39,6 @@
       <input id="nick" autocomplete="off" /><button>Ready</button>
     </form>
     <canvas id="screen" width="50" height="50"></canvas>
-    <script src="client.js"></script>
+    <script  src="client.js" type="module"></script>
 </body>
 </html>
